tweaks from review, refactor
