+++ homeassistant/components/braviatv/config_flow.py
@@ -1,4 +1,4 @@
-"""Adds config flow for Bravia TV integration."""
+"""Config flow to configure the Bravia TV integration."""
 from __future__ import annotations
 
 from contextlib import suppress
@@ -6,11 +6,11 @@ import ipaddress
 import re
 from typing import Any
 
-import aiohttp
-from pybravia import BraviaTV, BraviaTVError
+from aiohttp import CookieJar
+from pybravia import BraviaTV, BraviaTVError, BraviaTVNotSupported
 import voluptuous as vol
 
-from homeassistant import config_entries, exceptions
+from homeassistant import config_entries
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_MAC, CONF_PIN
 from homeassistant.core import callback
@@ -46,27 +46,16 @@ class BraviaTVConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
 
     def __init__(self) -> None:
         """Initialize."""
-        self.client: BraviaTV | None = None
-        self.host: str | None = None
+        self.client: BraviaTV
+        self.host: str
         self.title = ""
-        self.mac: str | None = None
+        self.mac: str
 
     async def init_device(self, pin: str) -> None:
         """Initialize Bravia TV device."""
-        assert self.client is not None
-
-        try:
-            await self.client.connect(
-                pin=pin, clientid=CLIENTID_PREFIX, nickname=NICKNAME
-            )
-        except BraviaTVError as err:
-            raise CannotConnect() from err
-
-        system_info = await self.client.get_system_info()
-        if not system_info:
-            raise ModelNotSupported()
-
+        await self.client.connect(pin=pin, clientid=CLIENTID_PREFIX, nickname=NICKNAME)
         await self.client.set_wol_mode(True)
+        system_info = await self.client.get_system_info()
 
         await self.async_set_unique_id(system_info[ATTR_CID].lower())
         self._abort_if_unique_id_configured()
@@ -90,7 +79,7 @@ class BraviaTVConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
             if host_valid(user_input[CONF_HOST]):
                 session = async_create_clientsession(
                     self.hass,
-                    cookie_jar=aiohttp.CookieJar(unsafe=True, quote_cookie=False),
+                    cookie_jar=CookieJar(unsafe=True, quote_cookie=False),
                 )
                 self.host = user_input[CONF_HOST]
                 self.client = BraviaTV(host=self.host, session=session)
@@ -114,19 +103,18 @@ class BraviaTVConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         if user_input is not None:
             try:
                 await self.init_device(user_input[CONF_PIN])
-            except CannotConnect:
-                errors["base"] = "cannot_connect"
-            except ModelNotSupported:
+            except BraviaTVNotSupported:
                 errors["base"] = "unsupported_model"
+            except BraviaTVError:
+                errors["base"] = "cannot_connect"
             else:
                 user_input[CONF_HOST] = self.host
                 user_input[CONF_MAC] = self.mac
                 return self.async_create_entry(title=self.title, data=user_input)
 
-        assert self.client is not None
-
-        pair = await self.client.pair(CLIENTID_PREFIX, NICKNAME)
-        if not pair:
+        try:
+            await self.client.pair(CLIENTID_PREFIX, NICKNAME)
+        except BraviaTVError:
             return self.async_abort(reason="no_ip_control")
 
         return self.async_show_form(
@@ -153,6 +141,7 @@ class BraviaTVOptionsFlowHandler(config_entries.OptionsFlow):
             self.config_entry.entry_id
         ]
 
+        await coordinator.async_update_sources()
         sources = coordinator.source_map.values()
         self.source_list = [item.get("title", "") for item in sources]
         return await self.async_step_user()
@@ -174,11 +163,3 @@ class BraviaTVOptionsFlowHandler(config_entries.OptionsFlow):
                 }
             ),
         )
-
-
-class CannotConnect(exceptions.HomeAssistantError):
-    """Error to indicate we cannot connect."""
-
-
-class ModelNotSupported(exceptions.HomeAssistantError):
-    """Error to indicate not supported model."""
