+++ requirements_all.txt
@@ -1402,7 +1402,7 @@ pyblackbird==0.5
 pybotvac==0.0.23
 
 # homeassistant.components.braviatv
-pybravia==0.1.7
+pybravia==0.2.0
 
 # homeassistant.components.nissan_leaf
 pycarwings2==2.13
