+++ requirements_test_all.txt
@@ -959,7 +959,7 @@ pyblackbird==0.5
 pybotvac==0.0.23
 
 # homeassistant.components.braviatv
-pybravia==0.1.7
+pybravia==0.2.0
 
 # homeassistant.components.cloudflare
 pycfdns==1.2.2
