+++ tests/components/braviatv/test_config_flow.py
@@ -1,7 +1,7 @@
 """Define tests for the Bravia TV config flow."""
 from unittest.mock import patch
 
-from pybravia import BraviaTVConnectionError
+from pybravia import BraviaTVConnectionError, BraviaTVNotSupported
 
 from homeassistant import data_entry_flow
 from homeassistant.components.braviatv.const import CONF_IGNORED_SOURCES, DOMAIN
@@ -53,9 +53,10 @@ async def test_user_invalid_host(hass):
 
 async def test_authorize_cannot_connect(hass):
     """Test that errors are shown when cannot connect to host at the authorize step."""
-    with patch("pybravia.BraviaTV.connect", side_effect=BraviaTVConnectionError), patch(
-        "pybravia.BraviaTV.pair", return_value=True
-    ):
+    with patch(
+        "pybravia.BraviaTV.connect",
+        side_effect=BraviaTVConnectionError,
+    ), patch("pybravia.BraviaTV.pair"):
         result = await hass.config_entries.flow.async_init(
             DOMAIN, context={"source": SOURCE_USER}, data={CONF_HOST: "bravia-host"}
         )
@@ -68,9 +69,10 @@ async def test_authorize_cannot_connect(hass):
 
 async def test_authorize_model_unsupported(hass):
     """Test that errors are shown when the TV is not supported at the authorize step."""
-    with patch("pybravia.BraviaTV.connect", return_value=True), patch(
-        "pybravia.BraviaTV.pair", return_value=True
-    ), patch("pybravia.BraviaTV.get_system_info", return_value={}):
+    with patch(
+        "pybravia.BraviaTV.connect",
+        side_effect=BraviaTVNotSupported,
+    ), patch("pybravia.BraviaTV.pair"):
         result = await hass.config_entries.flow.async_init(
             DOMAIN, context={"source": SOURCE_USER}, data={CONF_HOST: "10.10.10.12"}
         )
@@ -106,9 +108,12 @@ async def test_duplicate_error(hass):
     )
     config_entry.add_to_hass(hass)
 
-    with patch("pybravia.BraviaTV.connect", return_value=True), patch(
-        "pybravia.BraviaTV.pair", return_value=True
-    ), patch("pybravia.BraviaTV.get_system_info", return_value=BRAVIA_SYSTEM_INFO):
+    with patch("pybravia.BraviaTV.connect"), patch("pybravia.BraviaTV.pair"), patch(
+        "pybravia.BraviaTV.set_wol_mode"
+    ), patch(
+        "pybravia.BraviaTV.get_system_info",
+        return_value=BRAVIA_SYSTEM_INFO,
+    ):
 
         result = await hass.config_entries.flow.async_init(
             DOMAIN, context={"source": SOURCE_USER}, data={CONF_HOST: "bravia-host"}
@@ -123,10 +128,11 @@ async def test_duplicate_error(hass):
 
 async def test_create_entry(hass):
     """Test that the user step works."""
-    with patch("pybravia.BraviaTV.connect", return_value=True), patch(
-        "pybravia.BraviaTV.pair", return_value=True
+    with patch("pybravia.BraviaTV.connect"), patch("pybravia.BraviaTV.pair"), patch(
+        "pybravia.BraviaTV.set_wol_mode"
     ), patch(
-        "pybravia.BraviaTV.get_system_info", return_value=BRAVIA_SYSTEM_INFO
+        "pybravia.BraviaTV.get_system_info",
+        return_value=BRAVIA_SYSTEM_INFO,
     ), patch(
         "homeassistant.components.braviatv.async_setup_entry", return_value=True
     ):
@@ -154,10 +160,11 @@ async def test_create_entry(hass):
 
 async def test_create_entry_with_ipv6_address(hass):
     """Test that the user step works with device IPv6 address."""
-    with patch("pybravia.BraviaTV.connect", return_value=True), patch(
-        "pybravia.BraviaTV.pair", return_value=True
+    with patch("pybravia.BraviaTV.connect"), patch("pybravia.BraviaTV.pair"), patch(
+        "pybravia.BraviaTV.set_wol_mode"
     ), patch(
-        "pybravia.BraviaTV.get_system_info", return_value=BRAVIA_SYSTEM_INFO
+        "pybravia.BraviaTV.get_system_info",
+        return_value=BRAVIA_SYSTEM_INFO,
     ), patch(
         "homeassistant.components.braviatv.async_setup_entry", return_value=True
     ):
@@ -199,9 +206,27 @@ async def test_options_flow(hass):
     )
     config_entry.add_to_hass(hass)
 
-    with patch("pybravia.BraviaTV.connect", return_value=True), patch(
-        "pybravia.BraviaTV.get_power_status", return_value="active"
-    ), patch("pybravia.BraviaTV.get_external_status", return_value=BRAVIA_SOURCES):
+    # with patch("pybravia.BraviaTV.connect"), patch(
+    #     "pybravia.BraviaTV.get_power_status", return_value="active"
+    # ), patch("pybravia.BraviaTV.get_external_status", return_value=BRAVIA_SOURCES), patch(
+    #     "pybravia.BraviaTV.get_app_list", return_value={}
+    # ), patch(
+    #     "pybravia.BraviaTV.get_content_list_all", return_value={}
+    # ), patch(
+    #     "pybravia.BraviaTV.get_volume_info", return_value={}
+    # ), patch(
+    #     "pybravia.BraviaTV.get_playing_info", return_value={}
+    # ):
+    with patch("pybravia.BraviaTV.connect"), patch(
+        "pybravia.BraviaTV.get_power_status",
+        return_value="active",
+    ), patch(
+        "pybravia.BraviaTV.get_external_status",
+        return_value=BRAVIA_SOURCES,
+    ), patch(
+        "pybravia.BraviaTV.send_rest_req",
+        return_value={},
+    ):
         assert await hass.config_entries.async_setup(config_entry.entry_id)
         await hass.async_block_till_done()
